Removed use of win32 RECT struct here and there.  Replaced it with the Rect struct in misc.h
Added primatives to CBlitter.  Makes the name something of a misnomer, but it seems like the thing to do.

Exposed the CBlitter class to the interface, instead of making all kinds of pass-through functions in CPixelMatrix.  Removed CPixelMatrix::Blit, and all dependancies.

Tweaked graphics API.  gfxClipImage now accepts a Rect&, instead of a RECT (evil!)

Renamed winmaped's CGraphView::Rect to DrawRect, so as to not conflict with the Rect struct.